+++ src/__tests___/results/touchUI/clientlibs/hide.js
@@ -41,11 +41,7 @@
    * @returns {void} calls the function
    */
   function getFunction(str) {
-    const func = function () {
-      return str;
-    };
-
-    return func();
+    return new Function('return ' + str)();
   }
 
   /**